# src/docs/openapi.yaml
openapi: 3.0.0
...
info:
  title: E-Commerce REST API
  version: 1.0.0
  description: |
    Full-featured e-commerce backend API.

    ## Authentication
    Endpoints marked 🔒 require a JWT Bearer token.
    Obtain one via `POST /auth/login`, then set:
    `Authorization: Bearer <accessToken>`

    ## Token Lifecycle
    - **Access token** — expires in 15 minutes (in response body)
    - **Refresh token** — expires in 7 days (in `httpOnly` cookie)
    - Refresh via `POST /auth/refresh-token` (cookie sent automatically)

    ## Rate Limits
    - General: **100 req / 15 min** per IP
    - Auth: **10 req / 15 min** per IP
    - Uploads: **20 req / hr** per IP
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production

# =============================================================================
# Security
# =============================================================================
security:
  - bearerAuth: []

# =============================================================================
# Tags (sidebar groups)
# =============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    pageParam:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    limitParam:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    sortParam:
      in: query
      name: sort
      schema: { type: string, example: "-createdAt,price" }
    searchParam:
      in: query
      name: search
      schema: { type: string, example: "headphones" }
...
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        success:    { type: boolean, example: true }
        message:    { type: string,  example: "Success." }
        count:      { type: integer, example: 20 }
        pagination:
          type: object
          properties:
            total:       { type: integer, example: 150  }
            page:        { type: integer, example: 1    }
            limit:       { type: integer, example: 20   }
            totalPages:  { type: integer, example: 8    }
            hasNextPage: { type: boolean, example: true }
            hasPrevPage: { type: boolean, example: false }
        data: { type: array, items: {} }
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      example: "64abc1234567890def123456"
...
    CreateProductRequest:
      type: object
      required: [name, description, price, category, stock]
      properties:
        name:         { type: string,  minLength: 3,  maxLength: 200 }
        description:  { type: string,  minLength: 20, maxLength: 5000 }
        shortDesc:    { type: string,  maxLength: 200 }
        price:        { type: number,  minimum: 0 }
        comparePrice: { type: number,  minimum: 0 }
        category:     { $ref: "#/components/schemas/ObjectId" }
        stock:        { type: integer, minimum: 0 }
        sku:          { type: string  }
        brand:        { type: string  }
        tags:         { type: array,  items: { type: string } }
        isFeatured:   { type: boolean }
        isPublished:  { type: boolean }

    # ── Order ─────────────────────────────────────────────────────────────────
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: Get all products (paginated, filterable, searchable)
      security: []
      parameters:
        - { $ref: "#/components/parameters/pageParam" }
        - { $ref: "#/components/parameters/limitParam" }
        - { $ref: "#/components/parameters/sortParam" }
        - { $ref: "#/components/parameters/searchParam" }
        - in: query
          name: "price[gte]"
          schema: { type: number }
          description: Minimum price filter
        - in: query
          name: "price[lte]"
          schema: { type: number }
          description: Maximum price filter
        - in: query
          name: category
          schema: { type: string }
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedResponse" }
    post:
      tags: [Products]
      summary: 🔒 Admin/Seller — create a new product
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - { $ref: "#/components/schemas/CreateProductRequest" }
                - type: object
                  properties:
                    images:
                      type: array
                      items: { type: string, format: binary }
      responses:
        "201":
          description: Product created
...